feat(pj_media): visual markers (YOLO bboxes / annotations) over video
Implements Davide's documented multi-layer compositing architecture
(ARCHITECTURE.md §4.3, §5.4, §8) which had been deferred until
annotation test data was available. End-to-end trajectory: ObjectStore
bytes → SceneDecoder → ScenePipelineSource → CompositeMediaSource →
MediaViewerWidget (QRhi second pipeline for line primitives).

Core types and components (pj_media_core):
- scene_frame.h — POD types (Point2, ColorRGBA, PointsAnnotation,
  CircleAnnotation, TextAnnotation, ImageAnnotation, SceneFrame)
  mirroring datatypes_2D.md §8.
- media_frame.h — generalises MediaSource output to
  { optional<DecodedFrame> base; vector<SceneFrame> overlays; }.
  ImagePipelineSource, FileVideoSource, StreamingVideoSource updated
  to wrap their DecodedFrame in a MediaFrame.
- scene_decoder.{h,cpp} — ISceneDecoder interface + factory
  dispatching on MCAP schema name. Supported schemas:
    vision_msgs/msg/Detection2DArray  (CDR)
    yolo_msgs/msg/DetectionArray      (CDR)
    foxglove.ImageAnnotations         (Protobuf, hand-rolled wire parser)
  Each Detection2D maps to a LINE_LOOP PointsAnnotation of 4 corners.
  Colour comes from a stable 10-entry palette indexed by class_id %10
  (yolo_msgs has no colour field, so colour is derived from class).
- scene_pipeline_source.{h,cpp} — symmetric to ImagePipelineSource:
  setTimestamp() does latestAt() on ObjectStore, decodes, stages a
  pending SceneFrame; takeFrame() drains it into a MediaFrame overlay.
- composite_media_source.{h,cpp} — fans setTimestamp() across N layers,
  merges takeFrame() outputs into one MediaFrame (first base wins,
  overlays concatenated). Widget consumes a single MediaSource* with
  no knowledge of the layer model.

Rendering (pj_media_qt):
- Second QRhiGraphicsPipeline (Lines topology, alpha-blended) sharing
  the image pipeline's viewTransform UBO so markers track pan/zoom.
- New shaders scene_lines.{vert,frag} with frame_size + viewTransform
  uniforms. Vertex layout: vec2 pos + vec4 colour, 24-byte stride.
- expandToLineList() reserves vertex storage up-front and writes line
  segments via appendSegment(); only PointsAnnotation rendered (text
  and circles decoded but deferred — SDF font / instanced quads belong
  in a separate iteration).

Demos:
- mcap_image_viewer: McapTopicLoader generalisation (was image-only)
  consolidates two near-duplicate loaders behind a TopicMatchFn
  predicate. Discovers image + annotation topics in one pass and
  wires them into a CompositeMediaSource when both are present.
- multi_channel_viewer: same indexer treatment for parity.
- mcap_helpers.hpp: shared zstd magic-byte sniff + decompress used by
  both demos to handle rosbag2 compression_mode: MESSAGE payloads.

Bugs fixed during integration (captured in TECHNICAL_NOTES.md):
- mcap::ReadMessageOptions defaults to FileOrder, not LogTimeOrder.
  ObjectStore::pushLazy rejects out-of-order timestamps, so
  bag_compressed silently stopped at 132/140 messages. Set
  readOrder = LogTimeOrder explicitly in all loaders.
- rosbag2 compression_mode: MESSAGE produces zstd-compressed payloads
  (magic 0x28 b5 2f fd) that mcap does not transparently decompress.
  Demos now sniff and decompress before feeding the decoder.
- GLSL: a UBO field named imageSize is rewritten to _imageSize by
  SPIRV-Cross for GLSL ES100 (collides with the built-in imageSize()
  function), which broke QRhi name-based UBO mapping silently — the
  field stayed 0, NDC went NaN, all vertices were clipped. Renamed
  to frame_size.
- Marker pipeline must rebind viewport after the pipeline switch on
  some QRhi backends, and the SRB binds the UBO at slot 0 to BOTH
  vertex and fragment stages.

Datastore:
- object_store.hpp: one-line lock-order contract on entryTimestamps()
  documenting the shared lock held by EntryTimestampsView.
- object_store_test.cpp: three new concurrency tests
  (ConcurrentPushLazyAndAt, ConcurrentEntryTimestampsAndAt as the safe
  scoping pattern, MidFlightTeardown).
- OBJECT_STORE_DESIGN.md §3.4: matching subsection with safe-scoping
  example.

Docs (English, matching project convention):
- pj_media/docs/ARCHITECTURE.md §4.3, §5.4, §8.2 + component table:
  flipped from deferred to implemented with schema coverage table.
- pj_media/docs/datatypes_2D.md §8: implementation-status paragraph
  listing what renders today vs. what the framework supports but the
  renderer ignores.
- pj_media/docs/TECHNICAL_NOTES.md: four new lessons (GLSL imageSize
  pitfall, mcap readOrder default, rosbag2 message-mode zstd, marker
  viewport/UBO sharing).

Tests: 59/59 passing in build/ (RelWithDebInfo).
- New: scene_frame_test, scene_decoder_test, scene_pipeline_source_test,
  composite_media_source_test in pj_media_core.

nanocdr.hpp: explicit static_cast<uint32_t> on string/vector encode
length to silence -Wconversion under -Werror.

+++ pj_media/nanocdr/nanocdr.hpp
@@ -449,7 +449,7 @@ inline void Encoder::encode(const T& in) {
 }
 
 inline void Encoder::encode(const std::string& in) {
-  const uint32_t str_len = in.size();
+  const uint32_t str_len = static_cast<uint32_t>(in.size());
   encode(str_len);
   const auto prev_size = storage_->size();
   storage_->resize(prev_size + str_len);
@@ -458,7 +458,7 @@ inline void Encoder::encode(const std::string& in) {
 
 template <typename T, typename Allocator>
 inline void Encoder::encode(const std::vector<T, Allocator>& in) {
-  const uint32_t len = in.size();
+  const uint32_t len = static_cast<uint32_t>(in.size());
   encode(len);
   for (const auto& item : in) {
     encode(item);
